fix: navigate hash urls natively
Fixes https://github.com/webstudio-is/webstudio/issues/4830

Remix incorrectly navigates hashes and clear search params.
Here fixed this by using native `<a>` and tweaked the logic in builder.
