+++ packages/sdk-components-react-remix/src/rich-text-link.tsx
@@ -1,4 +1,1 @@
-import { RichTextLink as BaseLink } from "@webstudio-is/sdk-components-react";
-import { wrapLinkComponent } from "./shared/remix-link";
-
-export const RichTextLink = wrapLinkComponent(BaseLink);
+export { Link as RichTextLink } from "./link";
